Fix diffs in generated nuspec with nuget.core
There was a typo in file list that used Source
value for target attribute.

The ordering of metadata was different than
nuget.core and nuget.core never added
DevelopmentDependency if it was set to
false.

There are still some differences between a
nuspec generated through Manifest.Save
in Nuget.Core and this, but all of those
make it more consistent with the nuspec
that is put in the actual package so I'm
leaving them.

The primary difference is that NuGet.Core
used string types to represent the TFM in
groupable items so it would just pass through
whatever you gave the API to the nuspec.
The downside of this is that you might have
an invalid string and you wouldn't know that
until you packed (and cracked open the nuspec).
The new implementation parses the TFM
and writes a normalized version to the nuspec.
This is more wysiwyg so I'm leaving it.

Contained in branches: master#gitext://gotobranch/master
Contained in no tag