csc errors aren't getting written to summary message
The problem is when compile-csc is writing to Reporter.Output, it is using the original Console.Out instead of the intercepted Console.Out that was set when we invoke compile-csc in-proc.

The fix is to reset Reporter.Output and Error using the new Console.Out and Error, so when writing to Reporter.Output it gets interecepted.

Also fixing a performance optimization by overriding Write(string) in the LineNotificationTextWriter.