Adding all code pages for dotnet CLI.
Since we write to the console and create new processes and read from their stdout and stderr, we need to ensure we have all the code pages in our process.  This way we can encode/decode strings correctly when writing to our Console, and when reading from spawned process's output.

Fix #2486