"dotnet test project.json" fails in a directory with a test project.json in it
When the current directory contains a valid project.json, and the user just says `dotnet test project.json`, normalizing the path fails becuase we end up calling Path.GetFullPath with an empty string.

To fix this, if the directory of the file is empty, use the current directory.

Fixing this for all "dotnet XXX" commands.

Fix #2691