Add script to run and compare CLI perf tests
tests/Performance/run-perftests.py is a Python3 script that fetches all
of the dependencies needed to run the perf tests in test/Performance
locally. It can run the perf tests for a single dotnet.exe, or run two
dotnet.exe instances sequentially and compare the results between them.

Basic usage for a single test:
  run-perftests.py --name "runid" ...\dotnet.exe

Usage for a comparison run:
  run-perftests.py --name "runid" ...\dotnet.exe --base ...\dotnet.exe

For more detailed usage:
  run-perftests.py -h