Adding the symlink to dotnet binary to /usr/bin
This change removes adding the path through /etc/paths.d and
goes to adding a symlink to /usr/bin. The reason is that there is
a lot of users who are running into problems with zsh and oh-my-zsh has
a bug in the installer where they generate a .zshrc file with an
exploded $PATH at the time of install. This overrides the dynamic path
that is created with path_helper.

Fixes #3063, #2988, #1567