[DoNotMerge] EndToEnd test needs its own cache
This fixes #3322. There is an argument to be made that all tests should
do this, but as far as I can tell this and the HelloWorld test are the
only ones using a custom nuget.config which would prevent them from
inheriting using cli-deps.

I've also turned up verbosity of restore on this test so we can validate
the feeds it is pulling from in the future.

This can't be merged until a fix for #3315 goes in, so there isn't noise
in CI.