Allow selecting with either MIT or MSFT License
This change introduces a new configuration knob to the build
`InstallerLicenseType` which controls if the installers that we build are
licenses using the MIT or .NET Library license. The default is to continue
to use the .NET Library license.

For our zips/tar.gz's we will be moving to using the MIT license instead.
However, due to how these zips are built, that change will actually happen
in dotnet/core-setup (since the zips/tarballs from the host/host + shared
framework combo carry the license with them and that's just laid down on
disk when we pull artifacts from Azure).