Fix yum package conflict in Docker image
When running the CLI builds on the new build agents, the CentOS
build failed with the following error:

2017-01-26T20:39:30.0800810Z Step 4 : RUN yum -q -y install unzip libunwind gettext libcurl-devel openssl-devel zlib libicu-devel
2017-01-26T20:39:30.3020500Z  ---> Running in 97affb106685
2017-01-26T20:39:40.3872390Z [91mError: libselinux conflicts with fakesystemd-1-17.el7.centos.noarch

Swapping the the "fakesystemd" package with the real "systemd" package
fixed the issue, which was recommended in http://stackoverflow.com/a/36632668.

This issue hasn't been seen on the official builds because the build
agents in that pool have the Docker image cached, and they are using an
older version of the centos:7.1.1503 image which appears not to have
this issue.