Only restore crossgen for the current rid.
Today we download crossgen for all supported runtime ids. This
introduces a dependency on having builds for all supported platforms
in order to build the CLI on a single platform. This is problematic
for the build from source effort because we only build artifacts for a
specific rid (the RID of the host/target).

Update the CrossGen.Dependencies project to restore for only the RID
we are building for.