Suppress previous NuGet messages for clean command.
This commit suppresses messages that were previously generated during a restore
operation.  NuGet serializes the messages into `project.assets.json` during a
restore and the `ReportAssetsLogMessages` task reads and logs the messages.
This causes the clean command to print NuGet diagnostic messages even though a
restore operation does not occur.

The fix is to set the `EmitAssetsLogMessages` property to false for the clean
command, which prevents the `ReportAssetsLogMessages` target from executing.

Fixes #8027.