Remove unused Dia package
Extend CLI dependency updating
Extend CLI's dependency updating to include all repos.  At this time, only ASPNet won't work with this, as in prodcon it isn't specifying a branch in the manifest, which means no Latest_Packages.txt gets set.  Once this is fixed all upstream repos in CLI should be handled.
This will be combined with changes to the versions repo to update the subscriptions for 2.1.3xx CLI.