Better mock file system
This is a rewrite of the existing mock file system. The goal is to be as "real" as possible.
I've been working on it a bit every week for 2 months. The existing mock file system has many
behavior that is different than the real file system. Due to it is modeled by one big
directory. And it will give different result when the set up is different. I've investiaged
several hours in total for it, but I still cannot understand. This implementation uses a
tree with different types of object for Directory and File which is more accurate.