Avoid VT-100 output from run-build.sh
This is an attempt to work around/isolate the cause of intermittent
hangs in the CLI official build. Since the hang appears to be happening
when emitting a control code to STDOUT, let's try to set STDOUT up in
such a way that things mostly don't try to emit control codes to it.

Related:
- https://github.com/Microsoft/msbuild/issues/3563
- https://github.com/dotnet/corefx/issues/31559