NuGet prob nuget.config should start from the directory of the manifest file
The production change is rather small. However, nuget prob nuget config
and CLI prob manifest file is quite complex problem logic and it is even
worse when 2 are combined. The bug caused by it would be very nasty, as
in the test’s scenario. So, I added infrastructure to simulate this situation.