Add -DownloadDir parameter to dotnet-install
Our hosted build agents have multiple enlistments on them and we'd like to install the .NET Core SDK prior to building.  After the build we'd then delete the installation and re-install the version again for another build.  The current script deletes the downloaded zip file.

This change lets you specify a folder to use as a cache of downloaded installers so they can be used for installation again.